innitail commit
This is the landing pages that the user will come across. they will see a about us section of what the company is about then once they login or register they will then be taken to the more designed landing page that has all the listed events that are available. 
i was thinking of doing a business webinar or masterclass events where people can buy tickects of the business events that are coming up.